+++ docs/CHANGELOG.rst
@@ -26,6 +26,8 @@ Fixed
 - Output all detected input variants in ``mutations-table`` process
 - Change ``__`` separators in field names to ``_`` in ``general``
   descriptor schema
+- Change choice values of the field ``biomarkers_pdl1_tps_cat`` in
+  ``general`` descriptor schema
 
 ===================
 47.3.1 - 2022-10-18